requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
scipy
pyyaml
tqdm
matplotlib

# to run checks
pytest-cov
pytest
black
flake8
